requirements.txt
apache-airflow==2.7.3
pandas
praw
s3fs